% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency_precision}
\alias{adjacency_recall}
\title{Adjacency Precision and Recall}
\usage{
adjacency_precision(cgraph1, cgraph2)

adjacency_recall(cgraph1, cgraph2)
}
\arguments{
\item{cgraph1}{A causality graph}

\item{cgraph2}{A causality graph}
}
\value{
Both \code{adjacency_precision} and \code{adjacency_recall} return a
  numeric between 0 and 1 with the following exceptions:
  \itemize{
  \item \code{adjacency_precision} returns \code{NA} if \code{cgraph2}
    contains 0 adjacencies
  \item \code{adjacency_recall} returns \code{NA} if \code{cgraph1}
    contains 0 adjacencies
  }
}
\description{
\code{adjacency_precision} calculates the adjacency precison between two
causality graphs.
}
\details{
\code{adjacency_recall} calculates the adjacency recall between two
causality graphs.

TODO(arix) explain what they do
}
\examples{
TODO(arix)
}
\references{
Joseph D. Ramsey: “Scaling up Greedy Causal Search for Continuous Variables
  ”, 2015; \href{http://arxiv.org/abs/1507.07749}{arxiv:1507.07749 [cs.AI]}.
  TODO(arix) see what Spirtes calls these guys
  Spirtes et al: “Causation, Prediction, and Search.”, Mit Press,
  2001, p. 109.
}
\seealso{
Other graph comparison statistics:
\code{\link{arrowhead_precision}}, \code{\link{arrowhead_recall}},
  and \code{\link{shd}}
}
\author{
Alexander Rix
}
