% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency_precision}
\alias{adjacency_precision}
\alias{adjacency_recall}
\title{Adjacency Precision and Recall}
\usage{
adjacency_precision(cgraph1, cgraph2)

adjacency_recall(cgraph1, cgraph2)
}
\arguments{
\item{cgraph1}{A causality graph}

\item{cgraph2}{A causality graph}
}
\value{
Numeric between 0 and 1.

\code{adjacency_precision} returns \code{NA} if \code{cgprah2} has no
adjacencies.

Length one numeric between 0 and 1. If every node in \code{true_graph}
  is an orphan, \code{NA} is returned. A return value of 1 implies every
  adjacency in \code{true_graph} is also in \code{est_graph}
}
\description{
\code{adjacency_precision} calculates the adjacency precison between two
causality graphs.

\code{adjacency_recall} calculates the adjacency recall between two
causality graphs.
}
\details{
TODO(arix)

TODO(arix)
}
\examples{
TODO(arix)
}
\references{
Joseph D. Ramsey: “Scaling up Greedy Causal Search for Continuous Variables
  ”, 2015; \href{http://arxiv.org/abs/1507.07749}{arxiv:1507.07749 [cs.AI]}.

  Spirtes et al: “Causation, Prediction, and Search.”, Mit Press,
  2001, p. 109.
}
\seealso{
Other graph comparison statistics:
\code{\link{arrowhead_precision}}, \code{\link{arrowhead_recall}},
  and \code{\link{shd}}
}
\author{
Alexander Rix
}
