% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{aipw_rf_att}
\alias{aipw_rf_att}
\title{Estimate the average treatment effect on the treated (ATT) via augmented inverse propensity 
weighting.}
\usage{
aipw_rf_att(data, x, y, w, p, cf = TRUE, ...)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{x}{a list of character vectors specifying variables to be included in 
the model (columns in the data). If unspecified, then it is assumed to be all 
columns in the data besides y and w.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment status.}

\item{p}{a vector containing propensity score values.}

\item{cf}{logical; if TRUE then includes confidence interval on ATT.}

\item{...}{additional arguments to causal_forest.}
}
\value{
a list of ATT, 95 percent confidence interval upperbound and lowerbound or just ATT, depending on user input of \code{cf}.
}
\description{
Estimate the average treatment effect on the treated (ATT) via augmented inverse propensity 
weighting.
}
\details{
Computes an estimate of the ATT \eqn{\tau_T} with a augmented inverse propensity score weighted 
estimate on just the treated group (see aipw_ate for more details).
}
\examples{
data("lalonde")

# calculate propensity scores
p <- propensity_score(lalonde, y = "re78", w = "treat", model = "logistic")
att <- aipw_rf_att(data = lalonde, y = "re78", w = "treat", p = p, num.trees = 100, mtry = 3)

}
\references{
Rotnitzky, Andrea, James M. Robins, and Daniel O. Scharfstein. 1998. ``Semiparametric 
Regression for Repeated Outcomes with Nonignorable Nonresponse." \emph{Journal of the American Statistical Association}. 
Vol. 93, No. 444, Dec. pgs. 1321-1339.

Cao, Weihua, Anastasios A. Tsiatis, and Marie Davidian. 2009. ``Improving efficiency and robustness of the doubly 
robust estimator for a population mean with incomplete data". \emph{Biometrika}. Vol. 96(3). pgs. 723–734.
}
