% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead.R
\name{arrowhead_precision}
\alias{arrowhead_precision}
\alias{arrowhead_recall}
\title{Arrowhead Precision and Recall}
\usage{
arrowhead_precision(pdag1, pdag2)

arrowhead_recall(pdag1, pdag2)
}
\arguments{
\item{pdag1}{A causality PDAG}

\item{pdag2}{A causality PDAG}
}
\value{
Length one numeric between 0 and 1.

  \code{arrowhead_precision} returns \code{NA} if there are no oriented
  edges in \code{pdag2}.

\code{arrowhead_recall} returns \code{NA} if there are
  no oriented edges (arrows) in \code{pdag1}
}
\description{
\code{arrowhead_precision} calculates the arrowhead precision between two
causality graphs

\code{arrowhead_recall} calculates the arrowhead recall between two
causality graphs
}
\details{
\code{arrowhead_precision} counts the number of directed edges
  (\code{"-->"}) in \code{pdag1} and then counts how many directed edges in
  \code{pdag1} are also in \code{pdag2}. Then, the ratio is returned.

\code{arrowhead_recall} counts the number of directed edges
  \code{pdag1} and then counts how many directed edges in
  \code{pdag2} are in \code{pdag1}. Then, the ratio is returned. 1
  implies that every directed edge in \code{pdag1} are also in \code{pdag2}.
}
\examples{
TODO(arix)
}
\references{
Joseph D. Ramsey: “Scaling up Greedy Causal Search for Continuous Variables”,
 2015; \href{http://arxiv.org/abs/1507.07749}{arxiv:1507.07749[cs.AI]}.

  Spirtes et al. “Causation, Prediction, and Search.”, Mit Press,
  2001, p. 109.
}
\seealso{
Other graph comparison statistics:
  \code{\link{adjacency_precision}}, \code{\link{adjacency_recall}},
  and \code{\link{shd}}
  @export
}
\author{
Alexander Rix
}
