% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead.R
\name{arrowhead_recall}
\alias{arrowhead_recall}
\title{Determine how many arrows in graph 1 are in graph2.}
\usage{
arrowhead_recall(true_graph, est_graph)
}
\arguments{
\item{true_graph}{graph being treated as as the truth.}

\item{est_graph}{The graph being compared to the \code{true_graph}}
}
\value{
Numeric between 0 and 1;
\code{arrowhead_recall} returns \code{NA} if there are
  no orriented edges (arrows) in \code{true_graph}
}
\description{
\code{arrowhead_recall} calculates the arrowhead recall between a graph and an oracle
}
\details{
\code{arrowhead_recall} counts the number of directed edges (<--,
-->) in \code{true_graph} and then counts how many directed edges in
\code{est_graph} are in \code{true_graph}. Then, the ratio is returned. 1
implies that every directed edge in \code{true_graph} is also in
\code{est_graph}
}
\examples{
TODO(arix)
}
\references{
Joseph D. Ramsey: “Scaling up Greedy Causal Search for Continuous
  Variables”, 2015; \href{http://arxiv.org/abs/1507.07749}{arxiv:1507.07749 [cs.AI]}.
}
\seealso{
arrowhead_precision
}
