% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.R
\name{as.lavaan.formula}
\alias{as.lavaan.formula}
\title{Turn a causality graph into a lavaan formula}
\usage{
as.lavaan.formula(cgraph)
}
\arguments{
\item{cgraph}{A causality.graph to be turned into a lavaan formula}
}
\value{
a \code{lavaan} formula, or an error
}
\description{
Use this function to convert a causality.graph into a lavaan formula so you
can fit the model to the data and assess the quality of the model from the
graph.
}
\details{
currently, only DAGs are supported, though patterns and PDAGs with
  DAG extensions are accepted. Future work will inlcude PAGs, and including
  output from the FnFC family as a latent measurement model.
}
\examples{
# turn the dag from Sachs(2009) into a lavaan compatable formula
library(lavaan)
lavaan.forumula <- as.lavaan.formula(sachs.dag)
fit <- cfa(lavaan.formula, data = sachs.df)
}
