% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{assess_covariate_balance}
\alias{assess_covariate_balance}
\title{Assess balance in multivariate covariate distributions directly or by using 
propensity scores.}
\usage{
assess_covariate_balance(data, x, w, p = NULL, method = "Mahalanobis distance")
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{x}{a list of character vectors specifying the covariates.}

\item{w}{a character vector specifying the treatment variable. 
Treatment must be specified as 0 and 1 or TRUE and FALSE.}

\item{p}{a numeric vector of propensity scores. Only specified if \code{method = "NLPD"}.}

\item{method}{a character vector specifying the which method to use to assess balance 
(one of "Mahalanobis distance" or "NLPD")}
}
\value{
a numeric value for difference between the covariate distributions for treated and control groups.
}
\description{
Assess balance in multivariate covariate distributions directly or by using 
propensity scores.
}
\examples{
data(lalonde)

vars <- names(lalonde)
covariates <- vars[!vars \%in\% c("re78", "treat")]

assess_covariate_balance(lalonde, x = covariates, w = "treat")
p <- propensity_score(lalonde, y = "re78", w = "treat")

assess_covariate_balance(lalonde, x = covariates, w = "treat", p = p, method = "NLPD")
}
