% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{balance_plot}
\alias{balance_plot}
\title{Assess balance in covariate distribution graphically.}
\usage{
balance_plot(
  data,
  x,
  w,
  matched_indices = NULL,
  method = "Normalized difference",
  threshold = 0.2,
  colors = NULL,
  title = NULL
)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{x}{a list of character vectors specifying the covariates.}

\item{w}{a character vector describing the treatment variable.}

\item{matched_indices}{a data frame of two columns, one of treatment indices and one of 
matched (produced by \code{propensity_match}).}

\item{method}{a character vector specifying the method for which to assess balance in 
covariate distributions across treatment and control groups.}

\item{threshold}{the threshold to plot. Default is 0.2.}

\item{colors}{an optional vector of two specifying the colors for the control and
treatment groups, respectively.}

\item{title}{an optional plot title.}
}
\value{
a \code{ggplot} object.
}
\description{
Assess balance in covariate distributio and assess ability to adjust 
for differences in covariates by treatment status.
}
\examples{
data(lalonde)

p <- propensity_score(lalonde, y = "re78", w = "treat")
matched_indices <- propensity_match(lalonde, w = "treat", p = p, max_distance = .00001)

vars <- names(lalonde)
covariates <- vars[!vars \%in\% c("re78", "treat")]

balance_plot(lalonde, x = covariates, w = "treat", matched_indices = matched_indices)

}
