% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{causality-IO}
\alias{write_causality_graph}
\alias{read_causality_graph}
\title{Read and write causality graphs}
\usage{
write_causality_graph(file, graph)

read_causality_graph(file)
}
\arguments{
\item{file}{The file you wish to read/write \code{graph} from/to}

\item{graph}{The graph you wish to save.}
}
\value{
\code{read_causality_graph} returns an object of class
    "causality.graph". An error will be thrown if there does not exist a
     valid path to file, graph is not a graph, or the file you wish to read
     does not contain a graph in TETRAD format.
}
\description{
\code{write_causality_graph} writes a causality graph to file
}
\details{
\code{read_causality_graph} reads a causality graph from file
}
\note{
\code{file} will be output as a TETRAD compatible graph, and file must
be a TETRAD compatible graph if you wish to read it into R.
}
\examples{
\dontrun{write_causality_graph(file = "sachs", sachs.dag)}
# Will throw an error if PATH_TO_FILE is invalid
\dontrun{graph <-read_causality_graph(file = "PATH_TO_FILE")}
}
\author{
Alexander Rix
}
