\name{Causality-package}
\alias{Causality-package}
\alias{Causality}
\docType{package}
\title{
\packageTitle{Causality}
}
\description{
\packageDescription{Causality}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Causality}
\packageIndices{Causality}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{Causality}

Maintainer: \packageMaintainer{Causality}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
