% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.R
\name{CCF}
\alias{CCF}
\title{Cross Correlation Function}
\usage{
CCF(x, y, lag.max, ...)
}
\arguments{
\item{x}{A vector represents a time sequence}

\item{y}{A vector represents another time sequence}

\item{lag.max}{A numeric determines the max time lag}
}
\value{
$lag The sequence of time lag

$CCF The value of cross correlation function
}
\description{
Cross Correlation Function
}
\details{
The method of cross correlation function can be used to detect
the causality in the linear system.
}
\examples{
set.seed(42)
x<-runif(100)
y<-runif(100)
result<-CCF(x,y,lag.max=50)
plot(result)
}

