% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM.R
\name{CCM}
\alias{CCM}
\title{The original Convergent Cross Mapping}
\usage{
CCM(x, y, tau, E, LMN, ...)
}
\arguments{
\item{x}{A vector represent the time series}

\item{y}{A vector represent the time series}
}
\value{
A list of SugiX, SugiY, OriginX, OriginY, CorX, CorY
}
\description{
The original Convergent Cross Mapping
}
\details{
This method is proposed by the Sugihara, which is based on the Taken's Theorem.
}
\examples{
x<-c(1:100)
y<-c(100:1)
result<-CCM(x,y,tau=1,E=2)
}

