% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph-methods.R
\name{cgraph-methods}
\alias{cgraph-methods}
\alias{is.cyclic}
\alias{is.acyclic}
\alias{is.directed}
\alias{is.nonlatent}
\alias{is.latent}
\alias{parents}
\alias{children}
\title{Causality graph methods}
\usage{
is.cyclic(cgraph)

is.acyclic(graph)

is.directed(cgraph)

is.nonlatent(cgraph)

is.latent(cgraph)

parents(cgraph)

children(cgraph)
}
\arguments{
\item{graph}{a causality.graph}
}
\value{
\code{is.cyclic}, \code{is.acyclic}, \code{is.directed}, \code{is.nonlatent},
  and \code{is.latent} all return \code{TRUE} or \code{FALSE}
}
\description{
These are miscellaneous functions that can be applied any to causality.graph
object.
}
\details{
\itemize{
    \item
      \code{is.cyclic} and \code{is.acyclic} determine whether or not
      \code{graph} is cyclic/acyclic. It infers this from the class, or by
      using the sort method.
    \item
      \code{is.directed} determines if \code{graph} is directed, ie,
      the only edge type is \code{"-->"}
    \item
      \code{is.nonlatent} determines whether or not \code{graph}
      contains only nonlatent edge types (\code{"-->"}, \code{"---"}).
    \item
      \code{is.latent} determines whether or not \code{graph} contains
      only latent edge types
      (\code{"o->"}, \code{"o-o"}, \code{"<->"}, \code{"++>"}, \code{"~~>"}).
}

\code{is.directed} determines if \code{graph} is directed, ie,
  the only edge type is \code{"-->"}
}
