% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{cgraph}
\alias{cgraph}
\alias{is_valid_cgraph}
\alias{is.cgraph}
\alias{as.cgraph}
\title{Casuality Graphs}
\usage{
cgraph(nodes, edges, validate = T)

is_valid_cgraph(graph)

is.cgraph(graph)

as.cgraph(graph)
}
\arguments{
\item{nodes}{A character array of node names}

\item{edges}{A \eqn{m x 3} character matrix. Each row is an edge in the form
of (node1, node2, edgetype), with node1 and node2 being in nodes. Valid
edge types are listed below}

\item{validate}{logical number to determine whether or not to check to see
if the cgraph is valid before returning it. Default is \code{TRUE}}

\item{graph}{A graph to coerced or tested}
}
\value{
\code{cgraph} returns object of class "causality.graph", or an error
  if the graph is invalid.

\code{is_valid_cgraph} returns \code{TRUE} or \code{FALSE} depending
  on whether or not the input is valid
}
\description{
Create or test for objects of type "causality.graph"
}
\details{
A causality-graph consists of three things
  \itemize{
    \item nodes: a character vector of the nodes of the in the causal graph
    \item adjacencies a list of character vectors that contain the
      adjacencies of each node. This is calculated when a cgraph is created.
    \item edges a m x 3 character matrix which represents the edges in a
      causal graph in the form (from, to, edge). For example, if we are
      dealing with a causal graph regarding drug use and cancer, The edge
      "Smoking --> Cancer" would be stored as ("Smoking", "Cancer", "-->")
      in the edge matrix
  }
  The valid edges types for non latent variable model graphs
  (DAGs, PDAGs, Patterns) are:
  \itemize{
    \item \code{-->}
    \item \code{---}
  }
  And for latent variable models (PAGs, MAGs):
  \itemize{
    \item \code{o-o}
    \item \code{o->}
    \item \code{++>}
    \item \code{~~>}
    \item \code{<->}
  }

\code{is_valid_cgraph} checks to see if the input is a valid
"causality.graph." Specifally, it checks that there are no duplicate nodes,
as well as if the input is simple (no self-loops and is not a multi-graph).

\code{is.cgraph} tests whether or not an object has the class
  causality.graph
  @return \code{is.cgraph} returns \code{TRUE} or \code{FALSE}.

\code{as.cgraph} is an S3 generic that attempts to convert a not
causality.cgraph to a "causalality.graph". It currently supports turning "bn"
objects and r-causal objects to "causality.graphs".
}
\examples{
TODO(arix)
}
\references{
Spirtes, Peter, Clark N. Glymour, Richard Scheines, David Heckerman,
Christopher Meek, Gregory Cooper, and Thomas Richardson.
Causation, prediction, and search. MIT press, 2000.

Pearl, Judea. Causality. Cambridge university press, 2009.
}
\author{
Alexander Rix
}
