% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{check_propensity}
\alias{check_propensity}
\title{Check calibration and overlap assumptions via plot.}
\usage{
check_propensity(p, w, breaks = 20)
}
\arguments{
\item{p}{a vector of propensity scores.}

\item{w}{a vector of treatment indicators. 
Treatment must be specified as 0 and 1 or TRUE and FALSE.}

\item{breaks}{an integer number of breaks for the histogram, default to 20.}
}
\value{
a plot object that graphically tests overlap assumption and calibration.
}
\description{
Creates a Q-Q plot and a histogram to check 
calibration and overlap assumptions, respectively.
}
\examples{
data(lalonde)
p <- propensity_score(lalonde, y = "re78", w = "treat", model = "logistic", plot = FALSE)
w <- lalonde$treat
check_propensity(p = p, w = w)

}
