% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{dag}
\alias{dag}
\alias{is_valid_dag}
\alias{is.dag}
\alias{as.dag}
\alias{as.dag.default}
\alias{as.dag.causality.graph}
\alias{as.dag.causality.pdag}
\alias{as.dag.causality.pattern}
\alias{as.dag.causality.pag}
\title{Causality DAGS}
\usage{
dag(nodes, edges, validate = TRUE)

is_valid_dag(graph)

is.dag(graph)

as.dag(graph)

\method{as.dag}{default}(graph)

\method{as.dag}{causality.graph}(graph)

\method{as.dag}{causality.pdag}(graph)

\method{as.dag}{causality.pattern}(graph)

\method{as.dag}{causality.pag}(graph)
}
\arguments{
\item{nodes}{A character array of node names}

\item{edges}{A \eqn{m x 3} character matrix. Each row is an edge in the form
of (node1, node2, edgetype), with node1 and node2 being in nodes. Valid
edge types are listed below}

\item{validate}{logical value to determine whether or not to check to see
if the graph is valid before returning it. Default is \code{TRUE}}

\item{graph}{A graph to coerced or tested}
}
\value{
\code{dag} returns object of class "causality.dag", or an error
  if the graph is invalid (assuming \code{validate = TRUE}).

\code{is_valid_dag} returns \code{TRUE} or \code{FALSE} depending
  on whether or not the input is a valid "causality.dag".

\code{is.dag} returns \code{TRUE} or \code{FALSE}.
}
\description{
Create, test, or manipulate objects of type "causality.dag"
}
\details{
A causality DAG is a causality graph that is directed and acylic
  (hence the name DAG). DAGs are typically used to represent Bayesisan
  Networks and Structural Equation Models (SEMs).

\code{is_valid_dag} checks to see if the input is a valid
  "causality.dag". Specifically, it checks that the \code{graph} is directed
  and acyclic.

\code{is.dag} tests whether or not an object has the class
  "causality.dag"
}
\examples{
TODO
}
\references{
Spirtes et al. “Causation, Prediction, and Search.”, Mit Press,
  2001, p. 109.

 Spirtes P. Introduction to causal inference.
 Journal of Machine Learning Research. 2010;11(May):1643-62.

  Pearl, Judea. Causality. Cambridge university press, 2009.
}
\seealso{
Other causality classes: \code{\link{cgraph}}, \code{\link{pattern}}
}
\author{
Alexander Rix
}
