% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_to_pattern.R
\name{dag_to_pattern}
\alias{dag_to_pattern}
\title{Convert a DAG to a Pattern}
\usage{
dag_to_pattern(dag)
}
\arguments{
\item{dag}{An object of class dag that is to be converted}
}
\value{
A `causality` pattern. In the event that \code{dag} is not actually a
  dag, an error is thrown.
}
\description{
\code{dag_to_pattern} converts a causality dag to a pattern
}
\details{
The algorithm is due to Chickering(1995). See reference for details.
  The first step of the function is to perform a topoligical sort, which is
  only possible if \code{dag} is a dag.
}
\examples{
TODO(arix)
}
\references{
David Maxwell Chickering: “A Transformational Characterization of
  Equivalent Bayesian Network Structures”, 1995;
  \href{https://arxiv.org/abs/1302.4938}{arXiv:1302.4938 [cs.AI]} Cormen,
  Thomas H., et al. Introduction to Algorithms. The MIT Press, 2014.
}
