% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{double_selection_ate}
\alias{double_selection_ate}
\title{Estimate average treatment effect (ATE) with
double selection methodology.}
\usage{
double_selection_ate(data, x, y, w, nfolds = 5)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{x}{a list of character vectors specifying variables to be included in 
the model (columns in the data). If unspecified, then it is assumed to be all 
columns in the data besides y and w.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment status.}

\item{nfolds}{integer describing number of folds to use in k-fold cross-validation; preset to 5}
}
\value{
numeric estimate of the average treatment effect.
}
\description{
Estimate average treatment effect (ATE) with
double selection methodology.
}
\details{
Estimates the average treatment effect \eqn{\tau} using the methodology developed in
Belloni, Chernozhukov, and Hansen (2014), which they term the ``post-double-selection" method. 
The general procedure performed here is as follows: 
\enumerate{
  \item Predict the treatment \eqn{W_i} using the covariates \eqn{X_i} using lasso regression (where \eqn{\lambda} is tuned 
  using cross-validation). 
  Select the covariates that have non-zero coefficients in the lasso model. 
  \item Predict the outcome \eqn{Y_i} using the covariates \eqn{X_i} using lasso regression (where \eqn{\lambda} is tuned 
  using cross-validation). 
  Select the covariates that have non-zero coefficients in the lasso model. 
  \item Estimate the treatment effect \eqn{\tau} by the linear regression of \eqn{Y_i} on the treatment \eqn{W_i}
  and the union of the set of variables selected in the two covariate selection steps.
}
}
\examples{
data("lalonde")
ate <- double_selection_ate(data = lalonde, y = "re78", w = "treat")

}
\references{
Belloni, Alexandre, Victor Chernozhukov, and Christian Hansen. 2014. ``High-Dimensional Methods 
and Inference on Structural and Treatment Effects." \emph{Journal of Economic Perspective}.
Vol. 28, Num. 2, Spring 2014. pgs. 29–50.
\url{https://www.aeaweb.org/articles?id=10.1257/jep.28.2.29}
}
