\name{efflog-package}
\alias{efflog-package}
\alias{efflog}
\docType{package}
\title{Causal effects for loglinear models
}
\description{It calculates the causal effects
}
\details{
\tabular{ll}{
Package: \tab efflog\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-02\cr
License: \tab GPL (>=2)\cr
}}
\author{
Gloria Gheno <gloriagheno@libero.it>
}
\references{Gheno, G. (2015), The causal analysis in the loglinear model
}
\keyword{ package }
\seealso{
}
\examples{
}
