% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{general_match}
\alias{general_match}
\title{Performs general matching with exact and mahalanobis distnace.}
\usage{
general_match(data, w, max_distance = 0, replacement = TRUE, type = "exact")
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{w}{a character vector describing the treatment variable.}

\item{max_distance}{the maximum distance between propensity scores to form a match. 
Default is 0, as this is the distance for exact score matching}

\item{replacement}{logical; if FALSE then treated and control units 
are matched without replacement of control units (a control unit can
only be matched once).}

\item{type}{a string specifying which type of score matching to perform. 
Choices are 'exact', which is the default, and 'mahalanobis', which uses the Mahalanobis distance. 
If a different method is mentioned, then exact is performed by default.}
}
\value{
a dataframe with two columns, the first corresponding to indices for treated observations, 
and the second corresponding to indices for matched control observations.
}
\description{
Performs general matching with exact and mahalanobis distnace.
}
\examples{
data(lalonde)
general_match(lalonde, w = "treat", max_distance = 10, type = 'mahalanobis')
}
