% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.R
\name{ges}
\alias{ges}
\title{Greedy Equivalence Search (GES)}
\usage{
ges(df, score = c("bic", "bdue"), penalty = 1, sample.prior = 1,
  structure.prior = 1)
}
\arguments{
\item{df}{A data.frame with no missing values.}

\item{score}{The scoring function to use. Use BIC for continuous data and
BDeu for discrete.}

\item{penalty}{Tuning parameter for bic score. Cannot be less than 0;
less than 1 is probably a bad idea. Higher penalties will generate
sparser graphs. Defaults to 1, which corresponds to standard BIC.}

\item{sample.prior}{Second tuning parameter for BDeu score.}

\item{structure.prior}{First tuning parameter for BDeu score.}
}
\description{
GES is a score based causal discovery algorithm that outputs a pattern, a
graph that encodes the markov equilevence class of a set of DAGs. GES
contains score functions for continuous and discrete datasets. Mixed datasets
will have to be treated treated as continuous or discretized completly.
Other versions of ges support background knowledge, but this version does not.
}
\examples{
library(causality)
ges(ecoli.df, "bic", penalty = 2)
}
\references{
Chickering DM. Optimal structure identification with greedy search.
Journal of machine learning research. 2002;3(Nov):507-54.
}
\author{
Alexander Rix
}
