% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{ipw_ate}
\alias{ipw_ate}
\title{Etimate average treatment effect (ATE) with
inverse propensity score weighting.}
\usage{
ipw_ate(data, y, w, p, cf = T)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment status.}

\item{p}{a vector containing propensity score values.}

\item{cf}{logical; if TRUE then includes confidence interval on ATE.}
}
\value{
a list of ATE, 95 percent confidence interval upperbound and lowerbound or just ATE, depending on user input of \code{cf}
}
\description{
Etimate average treatment effect (ATE) with
inverse propensity score weighting.
}
\details{
Computes an estimate of the ATE \eqn{\tau} using propensity score weighting: 
\deqn{\tau = E \left[ \frac{Y_i W_i}{e(X_i)} - \frac{Y_i (1 - W_i)}{1 - e(X_i)} \right]}
where \eqn{e(X_i) = P(W_i = 1 | X_i = x)} is the propensity score.
}
\examples{
data("lalonde")

# calculate propensity scores
p <- propensity_score(data = lalonde, y = "re78", w = "treat", model = "logistic")
ate <- ipw_ate(data = lalonde, y = "re78", w = "treat", p = p)

}
\references{
Robins, James M., Andrea Rotnitzky and Lue Ping Zhao. 1994. "Estimation of Regression Coefficients 
When Some Regressors Are Not Always Observed." \emph{Journal of the American Statistical Association}. 
Vol. 89(427), Sep., pgs. 846-866. \url{https://doi.org/10.1080/01621459.1994.10476818}
Lunceford JK, Davidian M. 2004. ``Stratification and weighting via the propensity score in estimation of causal 
treatment effects: a comparative study." \emph{Statistics in Medicine}. Vol. 23(19), Aug., pgs. 2937–2960. 
\url{https://doi.org/10.1002/sim.1903}
}
