% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meek.R
\name{meek}
\alias{meek}
\title{Apply the Meek rules to a causality PDAG}
\usage{
meek(graph)
}
\arguments{
\item{graph}{a "causality.pdag"}
}
\value{
If a "causality.pdag is input, either either a pattern or dag is
  output. In the event that the pdag doesn't have a dag extension, garbage
  is output If a causality.dag or causality.pattern is input, \code{meek}
  just returns the input with no changes.
}
\description{
\code{meek} will turn a PDAG into a Pattern if the PDAG has a
valid dag extension
}
\details{
TODO(arix) -- Perhaps this function shouldn't be public
}
\note{
You are almostly certainly better off using the function
  \code{as.pattern}.
}
\examples{
TODO(arix)
}
\references{
Meek C. Causal inference and causal explanation with background knowledge.
  InProceedings of the Eleventh conference on Uncertainty in artificial
  intelligence 1995 Aug 18 (pp. 403-410). Morgan Kaufmann Publishers Inc..

Pearl, Judea. Causality. Cambridge university press, 2009.
}
