% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{naive_ate}
\alias{naive_ate}
\title{Estimate naive average treatment effect (ATE).}
\usage{
naive_ate(data, y, w, cf = TRUE)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment status.}

\item{cf}{logical; if TRUE then includes confidence interval on ATE.}
}
\value{
a list of ATE, 95 percent confidence interval upperbound and lowerbound or just ATE, depending on user input of \code{cf}
}
\description{
Estimate naive average treatment effect (ATE).
}
\details{
Computes a simple difference in means between the treatment group and the control group, 
\eqn{\tau = E[Y_i | W = 1] - E[Y_i | W = 0]}.
}
\examples{
data("lalonde")
ate <- naive_ate(data = lalonde, y = "re78", w = "treat")

}
