% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\alias{is_valid_pattern}
\alias{is.pattern}
\alias{as.pattern}
\alias{as.pattern.default}
\alias{as.pattern.causality.dag}
\alias{as.pattern.causality.pdag}
\alias{as.pattern.causality.pag}
\alias{as.pattern.causality.graph}
\title{Causality Patterns}
\usage{
pattern(nodes, edges, validate = TRUE)

is_valid_pattern(graph)

is.pattern(graph)

as.pattern(graph)

\method{as.pattern}{default}(graph)

\method{as.pattern}{causality.dag}(graph)

\method{as.pattern}{causality.pdag}(graph)

\method{as.pattern}{causality.pag}(graph)

\method{as.pattern}{causality.graph}(graph)
}
\arguments{
\item{edges}{A \eqn{m x 3} character matrix. Each row is an edge in the form
of (node1, node2, edgetype), with node1 and node2 being in nodes. Valid
edge types are listed below}

\item{validate}{logical value to determine whether or not to check to see
if the graph is valid before returning it. Default is \code{TRUE}}

\item{graph}{A graph to coerced or tested}
}
\value{
\code{dag} returns object of class "causality.pattern", or an error
  if the graph is invalid (assuming \code{validate = TRUE}).

\code{is_valid_pattern} returns \code{TRUE} or \code{FALSE} depending
  on whether or not the input is a valid "causality.pattern".

\code{is.pattern} returns \code{TRUE} or \code{FALSE}.
}
\description{
Create, test, or manipulate objects of type "causality.pattern"
}
\details{
TODO

\code{is_valid_pattern} checks to see if the input is a valid
  "causality.pattern". Specifically, it checks that the \code{graph} is
  nonlatent acyclic graph that is invarient under the \code{\link{meek}}
  rules (ie that it is a Completed PDAG -- AKA a pattern).

\code{is.pattern} tests whether or not an object has the class
  "causality.pattern"
}
\examples{
TODO
}
\references{
Spirtes et al. “Causation, Prediction, and Search.”, Mit Press,
  2001, p. 109.

 Spirtes P. Introduction to causal inference.
 Journal of Machine Learning Research. 2010;11(May):1643-62.

  Pearl, Judea. Causality. Cambridge university press, 2009.

  The algorithm used to convert dags to patterns is used by Chickering(1995)

  Chickering DM. A transformational characterization of equivalent Bayesian
  network structures. InProceedings of the Eleventh conference on
  Uncertainty in artificial intelligence 1995 Aug 18 (pp. 87-98).
  Morgan Kaufmann Publishers Inc..

  The algorithm by Dor and Tarsi is used to construct patterns from pdags

  Dor D, Tarsi M. A simple algorithm to construct a consistent extension of
  a partially oriented graph. Technicial Report R-185,
  Cognitive Systems Laboratory, UCLA. 1992 Oct 23.
}
\seealso{
Other causality classes: \code{\link{cgraph}}, \code{\link{dag}}
}
\author{
Alexander Rix
}
