% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdag.R
\name{pdag}
\alias{pdag}
\alias{is_valid_pdag}
\title{Partially Directed Acyclic Graphs}
\usage{
pdag(nodes, edges, validate = TRUE)

is_valid_pdag(graph)
}
\value{
\code{is_valid_pdag} returns \code{TRUE} or \code{FALSE} depending
  on whether or not the input is a valid "causality.pdag".
}
\description{
Partially Directed Acyclic Graphs
}
\details{
\code{is_valid_pdag} checks to see if the input is a valid
  "causality.pdag". Specifically, it checks that the \code{graph} is
  nonlatent and acyclic.
}
