% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_match}
\alias{propensity_match}
\title{Performs propensity score matching.}
\usage{
propensity_match(
  data,
  w,
  p,
  max_distance = 1,
  replacement = TRUE,
  type = "default"
)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{w}{a character vector describing the treatment variable.}

\item{p}{a numeric vector of propensity scores.}

\item{max_distance}{the maximum distance between propensity scores to form a match. 
Default is 1.}

\item{replacement}{logical; if FALSE then treated and control units 
are matched without replacement of control units (a control unit can
only be matched once).}

\item{type}{a string specifying which type of propensity score matching to perform. 
Choices are 'default', and 'linear'. Linear takes the logit of the propensity score before taking differences.
If not specified as linear, the default score matching will be used, which subtracts the scores as they are.}
}
\value{
a dataframe with two columns, the first corresponding to indices for treated observations, 
and the second corresponding to indices for matched control observations.
}
\description{
Performs propensity score matching.
}
\examples{
data(lalonde)

p <- propensity_score(lalonde, y = "re78", w = "treat")
propensity_match(lalonde, w = "treat", p = p, max_distance = 1)
}
