% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_score}
\alias{propensity_score}
\title{Calculate propensity scores.}
\usage{
propensity_score(
  data,
  x,
  y,
  w,
  subset = FALSE,
  model = "logistic",
  trim = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{x}{a list of character vectors specifying variables to be included in 
the model (columns in the data). If unspecified, then it is assumed to be all 
columns in the data besides y and w.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment variable. 
Treatment must be specified as 0 and 1 or TRUE and FALSE.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{model}{a character string naming the model used (one of "logistic", 
"lasso", or "causal forest").}

\item{trim}{a list of two numeric elements describing cutoffs at which to 
trim propensity scores.}

\item{plot}{logical; if TRUE then produces histogram of propensity scores to
check that the overlap assumption is satisfied and a Q-Q plot to check
the calibration.}

\item{...}{additional arguments to be passed to the model fitting functions.}
}
\value{
a vector of propensity scores where the indices of the vector correspond to the indices in \code{data}.
}
\description{
Function that calculates propensity scores with an option to
check calibration and overlap assumptions
}
\examples{
data(lalonde)
logit_propensities <- propensity_score(lalonde, y = "re78", w = "treat", model = "logistic")
lasso_propensities <- propensity_score(lalonde, y = "re78", w = "treat", model = "lasso")
cf_propensities <- propensity_score(lalonde, y = "re78", w = "treat", model = "causal forest")

}
\references{
Imbens, Guido, and Donald Rubin. \emph{Causal Inference for Statistics, Social, and Biomedical Sciences: An Introduction}. 
Cambridge University Press, 2015.
}
