% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{propensity_strat_ate}
\alias{propensity_strat_ate}
\title{Estimate average treatment effect (ATE) with
propensity score stratification.}
\usage{
propensity_strat_ate(data, y, w, p, n_strata = 4, min_obs = 1, model = "naive")
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment status.}

\item{p}{a vector containing propensity score values.}

\item{n_strata}{the number of strata to use when splitting propensity scores. Default: 4}

\item{min_obs}{Minimum number of observations per strata. 
If this is specified as a valid nontrivial number (greater than 1, less than non NA data size), 
then n_strata will be set to the minimum of the specified n_strata, and the n_strata such that min_obs is satisfied.}

\item{model}{a model to estimate ATE (one of "naive", "ipw", "weighted ols", or "aipw")}
}
\value{
numeric estimate of the average treatment effect.
}
\description{
Estimate average treatment effect (ATE) with
propensity score stratification.
}
\details{
Estimates average treatment effects across different groups of propensity scores specified
by the user (either with a minimum number of observations in a strata or number of strata). The user
can also specify the method of estimation.

The naive estimator is given by:

\eqn{\tau = E[Y_i | W = 1] - E[Y_i | W = 0]}.

The inverse propensity score weighting estimator is given by:

\deqn{\tau = E \left[ \frac{Y_i W_i}{e(X_i)} - \frac{Y_i (1 - W_i)}{1 - e(X_i)} \right]}
where \eqn{e(X_i) = P(W_i = 1 | X_i = x)} is the propensity score.

The weighted OLS model is a regression with weights given by:

\deqn{w_i = \frac{W_i}{e(X_i)} - \frac{1 - W_i}{1 - e(X_i)}}
where \eqn{e(X_i) = P(W_i = 1 | X_i = x)} is the propensity score. 

Finally, the augmented inverse propensity score weighted method is given by:

\deqn{\tau = E \left[  W_i \frac{Y_i-\tau(1,X_i)}{e(X_i)} + (1-W_i) \frac{Y_i-\tau(0,X_i)}{1-e(X_i)} + \tau(1,X_i) - \tau(0,X_i)\right]}
where \eqn{e(X_i) = P(W_i = 1 | X_i = x)} is the propensity score and \eqn{\tau(1, X_i)} and \eqn{\tau{0, X_i}} are estimated in
the first stage via OLS regression.

More details on each of these functions can be found in their individual estimators. The ATE is given by
the average of the estimates over the strata:

\eqn{\frac{1}{n}\sum_{i=1}^{n}\tau_i}
where \eqn{\tau_i} is the average treatment effect estimated in strata \eqn{i} and \eqn{n} is the number
of strata specified by the user (or calculated with the minimum number of observations per strata specified).
}
\examples{
data("lalonde")

# calculate propensity scores
p <- propensity_score(lalonde, y = "re78", w = "treat", model = "logistic")
ate <- propensity_strat_ate(data = lalonde, 
                            y = "re78", 
                            w = "treat",
                            p = p, 
                            n_strata = 4, 
                            min_obs = 1)

}
\references{
Rosenbaum, Paul, and Donald Rubin. 1984. ``Reducing Bias in Observational Studies Using Subclassification 
on the Propensity Score." \emph{Journal of the American Statistical Association}. Vol. 79(387), Feb. pgs 516-524.

Lunceford JK, Davidian M. 2004. ``Stratification and weighting via the propensity score in estimation of causal 
treatment effects: a comparative study." \emph{Statistics in Medicine}. Vol. 23(19), Aug., pgs. 2937–2960. 
\url{https://doi.org/10.1002/sim.1903}
}
