% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{rf_att}
\alias{rf_att}
\title{Uses random forests to naively estimate the average 
treatment effect on the treated (ATT) without weighting}
\usage{
rf_att(data, x, y, w, cf = TRUE, ...)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{x}{a list of character vectors specifying variables to be included in 
the model (columns in the data). If unspecified, then it is assumed to be all 
columns in the data besides y and w.}

\item{y}{a character vector specifying the response variable.}

\item{w}{a character vector specifying the treatment status.}

\item{cf}{logical; if TRUE then includes confidence interval on ATT.}

\item{...}{additional arguments to causal_forest.}
}
\value{
a list of ATT, 95 percent confidence interval upperbound and lowerbound or just ATT, depending on user input of \code{cf}.
}
\description{
Uses random forests to naively estimate the average 
treatment effect on the treated (ATT) without weighting
}
\details{
Computes an estimate of the ATT \eqn{\tau_T} with a naive estimate on just the treated group
(see naive_ate for more details).
}
\examples{
data("lalonde")
att <- rf_att(data = lalonde, y = "re78", w = "treat", num.trees = 100, mtry = 3)

}
\references{
Athey, Susan, Imbens, Guido, Pham, Thai, and Wager, Stefan. 2017. ``Estimating Average Treatment Effects: Supplementary Analyses and Remaining Challenges." 
\emph{The American Economic Review}, Vol. 107(5). pgs. 278–281.
url{www.jstor.org/stable/44250405}
}
