% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCCM.R
\name{runCCM}
\alias{runCCM}
\title{Run the Convergent Cross Mapping}
\usage{
runCCM(x, y, tau, E, L)
}
\arguments{
\item{x}{A vector represent the time series}

\item{y}{A vector represent the time series}

\item{tau}{The time interval of sample}

\item{E}{The embedding dimension}

\item{L}{Vector of computing points}
}
\value{
$L Vector of computing points

$CorX Vector of correlation X

$CorY Vector of correlation Y
}
\description{
Run the Convergent Cross Mapping
}
\details{
If x causes y, the correlation X will converge into a certain number.
}
\examples{
set.seed(43)
x<-runif(100)
set.seed(50)
y<-runif(100)
L<-seq(50,100,10)
tau<-1
E=2
dataCCM<-runCCM(x,y,tau = tau,E=E,L=L)
plot(dataCCM)
}

