% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagCCM.R
\name{sampleCCM}
\alias{sampleCCM}
\title{Generate the data structure for CCM}
\usage{
sampleCCM(x, y, lag, tag, E, ...)
}
\arguments{
\item{x}{A vector represents the time sequence}

\item{y}{A vector represents another time sequence}

\item{lag}{Time lag between x and y}

\item{tag}{A vector, which selects the dots of x and y to calculate the CCM}

\item{E+1}{is the embedded dimension of the manifold}
}
\value{
$dataX A matrix, which is manifold of X.
$dataY A vector, which is the original Y
}
\description{
Takes in two sequences, time lag, and row tag, E
}
\examples{
x<-c(1:100)
y<-c(100:1)
data<-sampleCCM(x,y,l=2,t = c(1:50),E=2)
}

