% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shd.R
\name{shd}
\alias{shd}
\title{Structural Hamming Distance}
\usage{
shd(x, y)
}
\arguments{
\item{x}{A causality.graph}

\item{y}{A causality.graph \code{x}}
}
\value{
Length one numeric between 0 and \eqn{n^2}, where n is the number of
  nodes. A distance of 0 implies that the patterns are the same
}
\description{
\code{shd} calculates the structural hamming distance (SHD) between two
patterns
}
\details{
\code{shd} takes in patterns and calculates the
  structural hamming distance between them as defined in Tsamardinos et
  al(2006). DAGs will also work, as they will be converted into patterns.
  Inputing PDAGs will work if the PDAGs admit DAG extensions. If (at least)
  one of them does not, an error is thrown.
}
\examples{
TODO(arix)
}
\references{
Tsamardinos I, Brown LE, Aliferis CF (2006). "The Max-Min
  Hill-Climbing Bayesian Network Structure Learning Algorithm". Machine
  Learning, 65(1), 31-78.
}
\seealso{
Other graph comparison statistics:
\code{\link{arrowhead_precision}}, \code{\link{arrowhead_recall}},
  \code{\link{adjacency_precision}}, and \code{\link{adjacency_recall}}
}
\author{
Alexander Rix
}
