% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_z_score.R
\name{test_fisher_independence}
\alias{test_fisher_independence}
\title{Test the Independence Between Continous variables.}
\usage{
test_fisher_independence(z_score, alpha)
}
\arguments{
\item{z_score}{Fisher z-score from (ideally) \code{fisher_z_score}, or elsewhere}

\item{alpha}{The signifigence level of the test, i.e., .05, .01, .001 etc}
}
\value{
\code{TRUE} if idependent or \code{FALSE} if not.
}
\description{
\code{test_fisher_independence} returns whether or not the \code{z_score} implies independence at the given signifigence level \code{alpha}
}
\details{
The test is H0: x, y are independent and H1: x,y are dependent.
}
\examples{
test_fisher_independence(5, .05)
test_fisher_independence(100,.001)
}
\seealso{
fisher_z_score
}
