% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{univariate_balance_plot}
\alias{univariate_balance_plot}
\title{Assess balance for a single covairate.}
\usage{
univariate_balance_plot(data, covariate, w, colors = NULL, title = NULL)
}
\arguments{
\item{data}{a dataframe object containing the variables and values.}

\item{covariate}{a character vectors specifying the covariate}

\item{w}{a character vector specifying the treatment variable.
Treatment must be specified as 0 and 1 or TRUE and FALSE.}

\item{colors}{an optional vector of two specifying the colors for the control and
treatment groups, respectively.}

\item{title}{an optional plot title.}
}
\value{
a \code{ggplot} object.
}
\description{
Assess balance for a single covairate.
}
\examples{
data(lalonde)

# histogram for continues variable: 
univariate_balance_plot(lalonde, "age", "treat")
# bar plot for discrete factor variable:
lalonde$hisp <- factor(lalonde$hisp)
univariate_balance_plot(lalonde, "hisp", "treat")

}
