% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead.R
\name{arrowhead_precision}
\alias{arrowhead_precision}
\alias{arrowhead_recall}
\title{Determine the percentage of arrows in a graph are in the true graph}
\usage{
arrowhead_precision(true_graph, est_graph)

arrowhead_recall(true_graph, est_graph)
}
\arguments{
\item{true_graph}{graph being treated as the truth.}

\item{est_graph}{The graph being compared to \code{true_graph}}
}
\value{
Length one numeric between 0 and 1.

  \code{arrowhead_precision} returns \code{NA} if there are no oriented
  edges in\code{est_graph}.

\code{arrowhead_recall} returns \code{NA} if there are
  no orriented edges (arrows) in \code{true_graph}
}
\description{
\code{arrowhead_precision} calculates the arrowhead precison between a graph
and an oracle.

\code{arrowhead_recall} calculates the arrowhead recall between a graph and an oracle
}
\details{
\code{arrowhead_precision} counts the number of directed edges
  ("node1", "node2", \code{"-->"}) in \code{est_graph} and then counts how many
  directed edges in \code{est_graph} are also in \code{true_graph}. Then, the
  ratio is returned. 1 implies that every directed edge
  in \code{est_graph} is also in \code{true_graph}, while 0 implies that no
  directed edge in \code{est_graph} is in \code{true_graph}

\code{arrowhead_recall} counts the number of directed edges (<--,
-->) in \code{true_graph} and then counts how many directed edges in
\code{est_graph} are in \code{true_graph}. Then, the ratio is returned. 1
implies that every directed edge in \code{true_graph} is also in
\code{est_graph}
}
\examples{
TODO(arix)
}
\references{
Joseph D. Ramsey: “Scaling up Greedy Causal Search for Continuous
  Variables”, 2015; \href{http://arxiv.org/abs/1507.07749}{arxiv:1507.07749
  [cs.AI]}.

  Spirtes et al. “Causation, Prediction, and Search.”, Mit Press,
  2001, p. 109.
}
\seealso{
Other graph comparison statistics: \code{\link{adjacency_precision}},
  \code{\link{adjacency_recall}}, \code{\link{shd}}
}
