% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead.R
\name{arrowhead_precision}
\alias{arrowhead_precision}
\title{Determine the percentage of arrows in a graph are in the true graph}
\usage{
arrowhead_precision(true_graph, est_graph)
}
\arguments{
\item{true_graph}{graph being treated as the truth.}

\item{est_graph}{The graph being compared to \code{true_graph}}
}
\value{
Length one numeric between 0 and 1. If there are no oriented edges in
  \code{est_graph}, \code{NA} is returned. 1 implies that every directed edge
  in \code{est_graph} is also in \code{true_graph}, while 0 implies that no
  directed ege in \code{est_graph} is in \code{true_graph}
}
\description{
\code{arrowhead_precision} calculates the arrowhead precison between a graph
and an oracle.
}
\details{
\code{arrowhead_precision} counts the number of directed edges
  (\code{"<--"},\code{"-->"}) in \code{est_graph} and then counts how many
  directed edges in \code{est_graph} are also in \code{true_graph}. Then, the
  ratio is returned.
}
\examples{
TODO(arix)
}
\references{
Joseph D. Ramsey: “Scaling up Greedy Causal Search for Continuous
  Variables”, 2015; \url{http://arxiv.org/abs/1507.07749} [cs.AI].
}
\seealso{
\code{\link{arrowhead_recall}}
}
