% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgraph.R
\name{as.cgraph}
\alias{as.cgraph}
\alias{as.cgraph.causality.graph}
\alias{as.cgraph.default}
\alias{as.cgraph.bn}
\title{Coerce a graph to a Causality Graph}
\usage{
as.cgraph(graph)

\method{as.cgraph}{causality.graph}(graph)

\method{as.cgraph}{default}(graph)

\method{as.cgraph}{bn}(graph)
}
\arguments{
\item{graph}{A (non causality.graph) graph you'd like to attempt to convert
into a causality.graph}
}
\value{
\code{as.cgraph} returns a causality graph object, or throws an error
}
\description{
Coerce a graph to a Causality Graph
}
\details{
\code{as.cgraph} is an S3 generic that attempts to convert a not
  causality.cgraph to a "causalality.graph". It currently supports turning
  "bn" objects and r-causal objects to "causality.graphs".
}
\seealso{
See \code{\link{cgraph}} for the documentation of object
}
