% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{chickering}
\alias{chickering}
\title{Turn a causality DAG into a pattern.}
\usage{
chickering(graph)
}
\arguments{
\item{graph}{a causality dag.}
}
\value{
a causality.pattern
}
\description{
\code{chickering} converts dags into their markov equivalence class. This
function exists mostly to provide documentation on how causality dags are
converted to patterns (or pdags). Using as.pattern instead is recommended.
}
\references{
Chickering DM. A transformational characterization of equivalent Bayesian
   network structures. InProceedings of the Eleventh conference on
   Uncertainty  in artificial intelligence 1995 Aug 18 (pp. 87-98). Morgan
   Kaufmann Publishers Inc. \href{http:///arxiv.org/abs/1302.4938}{arXiv:1302.4938 [cs.AI]}.
}
