% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meek.R
\name{meek}
\alias{meek}
\title{Apply the Meek rules to a causality PDAG}
\usage{
meek(graph)
}
\arguments{
\item{graph}{a "causality.pdag"}
}
\value{
\code{meek} either returns a causality.dag, causality.pattern,
  or causality.pdag.
}
\description{
\code{meek} will maximally orient the given causality.pdag, possibly turning
it into a pattern or dag.
}
\details{
\code{meek} applies the meek rules to the given graph. Applying the
  rules results in an a maximally oriented pdag. Depending on the exact
  structure of graph, the output graph is either a dag, pattern, or pdag.
}
\note{
You are almostly certainly better off using the function
  \code{as.pattern}.
}
\examples{
# Create a causality.pdag
nodes <- c("X1", "X2", "X3", "X4", "X5")
edges <- matrix(c("X1", "X5", "---",
                  "X1", "X2", "-->",
                  "X3", "X2", "-->",
                  "X3", "X4", "---"), ncol = 3, byrow = T)
graph <- cgraph(nodes, edges)
# Applying the meek rules will turn this graph into a pattern
meek(graph)
}
\references{
Meek C. Causal inference and causal explanation with background knowledge.
  InProceedings of the Eleventh conference on Uncertainty in artificial
  intelligence 1995 Aug 18 (pp. 403-410). Morgan Kaufmann Publishers Inc..

Pearl, Judea. Causality. Cambridge university press, 2009.
}
