% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shd.R
\name{shd}
\alias{shd}
\title{Calculate the Structural Hamming Distance between two PDAGS}
\usage{
shd(pdag1, pdag2)
}
\arguments{
\item{pdag1}{A causality pdag being used as the "true" pdag}

\item{pdag2}{A causality pdag that is being compared to \code{pdag1}}
}
\value{
Length one numeric between 0 and \eqn{n^2}, where n is the number of
  nodes.
}
\description{
\code{adjacency_precision} calculates the adjacency precison between a graph
and an oracle.
}
\details{
\code{shd} takes in objects of class pdag and calculates the
  structural hamming distance between them as defined in Tsamardinos et
  al(2006). Note that patterns are a subset of dpags, so inputing these
  objects works as well. If the event the at least 1 of the inputs is a dag,
  shd returns an error and states that you need to run \code{dag_to_pattern}
  to coerce the dag to a pattern first.
}
\examples{
TODO(arix)
}
\references{
Tsamardinos I, Brown LE, Aliferis CF (2006). "The Max-Min
  Hill-Climbing Bayesian Network Structure Learning Algorithm". Machine
  Learning, 65(1), 31-78.
}
\seealso{
\code{\link{dag_to_pattern}}
}
